+++ src/main/java/jpa/domain/BaseEntity.java
@@ -1,0 +1,24 @@
+package jpa.domain;
+
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import javax.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
+
+@MappedSuperclass
+public abstract class BaseEntity {
+  @CreationTimestamp
+  private LocalDateTime createdDate;
+
+  @UpdateTimestamp
+  private LocalDateTime modifiedDate;
+
+  public LocalDateTime getCreatedDate() {
+    return createdDate;
+  }
+
+  public LocalDateTime getModifiedDate() {
+    return modifiedDate;
+  }
+}
