+++ src/main/java/jpa/domain/Station.java
@@ -1,31 +1,26 @@
 package jpa.domain;
 
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
 import javax.persistence.*;
-import java.time.LocalDateTime;
 
-@Entity(name = "station")
-public class Station {
+@Entity
+public class Station extends BaseEntity {
 
   @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY) // auto increment
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
   @Column(nullable = false, unique = true)
   private String name;
 
-  @CreationTimestamp
-  private LocalDateTime createdDate;
-
-  @UpdateTimestamp
-  private LocalDateTime modifiedDate;
-
   protected Station() {
+
   }
 
   public Station(final String name) {
+    this.name = name;
+  }
+
+  public void changeName(String name) {
     this.name = name;
   }
 
@@ -35,17 +30,5 @@ public class Station {
 
   public String getName() {
     return name;
-  }
-
-  public LocalDateTime getCreatedDate() {
-    return createdDate;
-  }
-
-  public LocalDateTime getModifiedDate() {
-    return modifiedDate;
-  }
-
-  public void changeName(String name) {
-    this.name = name;
   }
 }
