+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.show-sql=true
-#spring.jpa.hibernate.ddl-auto=
+spring.jpa.hibernate.ddl-auto=create
